+++ CHANGELOG.md
@@ -110,6 +110,8 @@ For instructions on upgrading to newer versions, visit
 
 ### Resolved Issues
 
+* \#2393 Ensure `inverse_of` is respected when using polymorphic relations.
+
 * \#2388 Map/reduce properly uses `sort` instead of `orderby` in the execution
   of the command. (Alex Tsibulya)